requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
asyncssh>=2.14.0
python-multipart>=0.0.6

# Development
httpx>=0.25.0  # for testing
